componetized dashboard and styled app tiles
