+++ client/src/components/AuthenticationGuard.jsx
@@ -1,7 +1,10 @@
 import { withAuthenticationRequired } from "@auth0/auth0-react";
 import PageLoader from "./PageLoader";
 
-const AuthenticationGuard = ({ component }) => {
+const AuthenticationGuard = (props) => {
+  const component = props.component;
+  const searchQuery = props.searchQuery;
+
   const Component = withAuthenticationRequired(component, {
     onRedirecting: () => (
       <div className="page-layout">
@@ -10,7 +13,7 @@ const AuthenticationGuard = ({ component }) => {
     ),
   });
 
-  return <Component />;
+  return <Component searchQuery={searchQuery} />;
 };
 
 export default AuthenticationGuard;
