--- client/src/reusable/StyledComponents.jsx
@@ -1,1 +1,0 @@
-import styled from "styled-components";
